More changes

+++ demo_modl_singlechannel/utils/flare_utils.py
@@ -199,7 +199,7 @@ def conjgrad(x, b, Aop_fun, max_iter=10, l2lam=0., eps=1e-4, verbose=True):
         if verbose:
             print('{i}: {rsnew}'.format(i=i, rsnew=dbp.utils.itemize(torch.sqrt(rsnew))))
 
-        if rsnew.float().max() < eps:
+        if rsnew.max() < eps:
             break
 
         if l2lam > 0:
